+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 - Fixed the construction of the block hash table (#1506)
 - Fixed a bug in the block stack table (#1511) (#1512)
 - Fixed the construction of the chiplets virtual table (#1514)
-- Fixed the construction of the chiplets bus (#1516)
+- Fixed the construction of the chiplets bus (#1516) (#1525)
 
 #### Fixes